+++ Backend/Model/Imagem.cs
@@ -14,7 +14,7 @@ namespace WinstonChurchill.Backend.Model
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("ID")]
-        public decimal ID { get; set; }
+        public int ID { get; set; }
 
         [Column("NomeArquivo")]
         public string NomeArquivo { get; set; }
@@ -38,7 +38,7 @@ namespace WinstonChurchill.Backend.Model
         #region Foreign Keys
 
         [Column("UsuarioID")]
-        public decimal UsuarioID { get; set; }
+        public int UsuarioID { get; set; }
 
         [ForeignKey("UsuarioID")]
         public Usuario Usuario { get; set; }
